requirements.txt
pyusb==1.0.0
SQLAlchemy==1.1.5